Add postgres db, test passing
