+++ crates/broker/src/submitter.rs
@@ -406,7 +406,7 @@ where
 mod tests {
     use super::*;
     use crate::{
-        db::SqliteDb,
+        db::{DBPoolManager, SqliteDb},
         now_timestamp,
         provers::{encode_input, MockProver},
         AggregationState, Batch, BatchStatus, Order, OrderStatus,
